change!: move `fetch::Command` into the crate root.
This represents much better on how it is actually used, which is for
validation of any Command invoked through the protocol independenty.
