+++ git-protocol/src/command/tests.rs
@@ -8,10 +8,8 @@ mod v1 {
     const GITHUB_CAPABILITIES: &str = "multi_ack thin-pack side-band ofs-delta shallow deepen-since deepen-not deepen-relative no-progress include-tag allow-tip-sha1-in-want allow-reachable-sha1-in-want no-done symref=HEAD:refs/heads/main filter agent=git/github-gdf51a71f0236";
     mod fetch {
         mod default_features {
-            use crate::fetch::{
-                tests::command::v1::{capabilities, GITHUB_CAPABILITIES},
-                Command,
-            };
+            use crate::command::tests::v1::{capabilities, GITHUB_CAPABILITIES};
+            use crate::Command;
 
             #[test]
             fn it_chooses_the_best_multi_ack_and_sideband() {
@@ -59,7 +57,8 @@ mod v2 {
 
     mod fetch {
         mod default_features {
-            use crate::fetch::{tests::command::v2::capabilities, Command};
+            use crate::command::tests::v2::capabilities;
+            use crate::Command;
 
             #[test]
             fn all_features() {
@@ -79,7 +78,8 @@ mod v2 {
         mod initial_arguments {
             use bstr::ByteSlice;
 
-            use crate::fetch::{tests::command::v2::capabilities, Command};
+            use crate::command::tests::v2::capabilities;
+            use crate::Command;
 
             #[test]
             fn for_all_features() {
@@ -99,7 +99,8 @@ mod v2 {
 
     mod ls_refs {
         mod default_features {
-            use crate::fetch::{tests::command::v2::capabilities, Command};
+            use crate::command::tests::v2::capabilities;
+            use crate::Command;
 
             #[test]
             fn default_as_there_are_no_features() {
@@ -116,7 +117,8 @@ mod v2 {
         mod validate {
             use bstr::ByteSlice;
 
-            use crate::fetch::{tests::command::v2::capabilities, Command};
+            use crate::command::tests::v2::capabilities;
+            use crate::Command;
 
             #[test]
             fn ref_prefixes_can_always_be_used() {
