+++ git-protocol/src/fetch/arguments/blocking_io.rs
@@ -2,7 +2,8 @@ use std::io::Write;
 
 use git_transport::{client, client::TransportV2Ext};
 
-use crate::fetch::{Arguments, Command};
+use crate::fetch::Arguments;
+use crate::Command;
 
 impl Arguments {
     /// Send fetch arguments to the server, and indicate this is the end of negotiations only if `add_done_argument` is present.
