+++ git-protocol/src/fetch/tests/mod.rs
@@ -1,5 +1,4 @@
 #[cfg(any(feature = "async-client", feature = "blocking-client"))]
 mod arguments;
-mod command;
 #[cfg(any(feature = "blocking-client", feature = "async-client"))]
 mod refs;
